[risk=low][RW-8314] Don't crash on missing recent-resources references (#6660)
* isValidResource() now also applies to data resources
* tests for isValidResource and testGetUserRecentResource for deleted resources

+++ api/src/main/java/org/pmiops/workbench/db/model/DbUserRecentlyModifiedResource.java
@@ -21,7 +21,7 @@ public class DbUserRecentlyModifiedResource {
     COHORT_REVIEW
   }
 
-  private int id;
+  private Long id;
   private Long userId;
   private Long workspaceId;
   private DbUserRecentlyModifiedResourceType resourceTypeEnum;
@@ -45,11 +45,11 @@ public class DbUserRecentlyModifiedResource {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  public int getId() {
+  public Long getId() {
     return id;
   }
 
-  public DbUserRecentlyModifiedResource setId(int id) {
+  public DbUserRecentlyModifiedResource setId(Long id) {
     this.id = id;
     return this;
   }
